# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users/me": {
      "get": {
        "operationId": "UserController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserProfileDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the current user's profile",
        "tags": [
          "User"
        ]
      },
      "put": {
        "operationId": "UserController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update the current user's profile",
        "tags": [
          "User"
        ]
      }
    },
...
    "/api/transfer": {
      "post": {
        "operationId": "TransferController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransferDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/CreateTransferResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send money to another account (P2P). Atomic: balance check, paired transactions, and a notification to the receiver are all committed together or rolled back.",
        "tags": [
          "Transfer"
        ]
      }
    },
    "/api/transfer/history": {
      "get": {
        "operationId": "TransferController_history",
        "parameters": [
          {
            "name": "direction",
            "required": false,
            "in": "query",
            "description": "Filter by direction (omit to include both)",
            "schema": {
              "type": "string",
              "enum": [
                "sent",
                "received"
              ]
            }
          },
          {
            "name": "start",
            "required": false,
            "in": "query",
            "description": "Start date (inclusive)",
            "schema": {
              "example": "2026-05-01",
              "type": "string"
            }
          },
          {
            "name": "end",
            "required": false,
            "in": "query",
            "description": "End date (inclusive)",
            "schema": {
              "example": "2026-05-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of transfer history items (newest first)",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/TransferHistoryItemDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List transfers where the current user is the sender, receiver, or both.",
        "tags": [
          "Transfer"
        ]
      }
    },
...
  },
  "info": {
    "title": "Sonsam Saving API",
    "description": "Personal Finance Tracker — REST API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 200,
            "description": "HTTP status code"
          },
          "message": {
            "type": "string",
            "example": "Success",
            "description": "Human-readable message"
          },
          "data": {
            "type": "object",
            "description": "Response payload (null when there is no body)"
          }
        },
        "required": [
          "statusCode",
          "message",
          "data"
        ]
      },
...
      "UserProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "description": "User ID (BigInt serialized as string)"
          },
          "fullName": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "phone": {
            "type": "object",
            "example": "012345678",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-04-29T04:43:29.250Z"
          }
        },
        "required": [
          "id",
          "fullName",
          "email",
          "createdAt"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "example": "John Updated"
          },
          "phone": {
            "type": "string",
            "example": "098765432"
          }
        }
      },
...
      "TransferReceiverDto": {
        "type": "object",
        "properties": {
          "accountNumber": {
            "type": "string",
            "example": "ACC-20260429-00002"
          },
          "name": {
            "type": "string",
            "example": "Jane Doe"
          }
        },
        "required": [
          "accountNumber",
          "name"
        ]
      },
      "CreateTransferResponseDto": {
        "type": "object",
        "properties": {
          "transferId": {
            "type": "string",
            "example": "5"
          },
          "amount": {
            "type": "string",
            "example": "20.00",
            "description": "Decimal serialized as string"
          },
          "receiver": {
            "$ref": "#/components/schemas/TransferReceiverDto"
          },
          "newBalance": {
            "type": "string",
            "example": "230.00",
            "description": "Sender's balance after the transfer"
          }
        },
        "required": [
          "transferId",
          "amount",
          "receiver",
          "newBalance"
        ]
      },
      "CreateTransferDto": {
        "type": "object",
        "properties": {
          "receiverAccountNumber": {
            "type": "string",
            "example": "ACC-20260429-00002",
            "description": "Receiver account number in format ACC-YYYYMMDD-NNNNN"
          },
          "amount": {
            "type": "number",
            "example": 20,
            "description": "Positive amount, max 2 decimals"
          },
          "note": {
            "type": "string",
            "example": "Treat for coffee",
            "maxLength": 500
          }
        },
        "required": [
          "receiverAccountNumber",
          "amount"
        ]
      },
      "TransferCounterpartDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "accountNumber": {
            "type": "string",
            "example": "ACC-20260429-00002"
          }
        },
        "required": [
          "name",
          "accountNumber"
        ]
      },
      "TransferHistoryItemDto": {
        "type": "object",
        "properties": {
          "transferId": {
            "type": "string",
            "example": "5"
          },
          "direction": {
            "type": "string",
            "enum": [
              "sent",
              "received"
            ],
            "example": "sent",
            "description": "sent — current user is the sender; received — current user is the receiver"
          },
          "amount": {
            "type": "string",
            "example": "20.00"
          },
          "note": {
            "type": "object",
            "example": "Treat for coffee",
            "nullable": true
          },
          "counterpart": {
            "$ref": "#/components/schemas/TransferCounterpartDto"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-05-10T14:00:00.000Z"
          }
        },
        "required": [
          "transferId",
          "direction",
          "amount",
          "counterpart",
          "createdAt"
        ]
      },
...
    }
  }
}